refactor(config): add read file
